+++ src/main/java/com/toondeboer/utils/Solution.java
@@ -11,15 +11,15 @@ public abstract class Solution {
         String fileName = "day" + day + ".txt";
         String input = InputReader.readInput(fileName);
 
-        int solutionPart1 = solvePart1(input);
-        int solutionPart2 = solvePart2(input);
+        long solutionPart1 = solvePart1(input);
+        long solutionPart2 = solvePart2(input);
 
         System.out.println("Day " + day
                 + "\n\tPart 1: " + solutionPart1
                 + "\n\tPart 2: " + solutionPart2);
     }
 
-    public abstract int solvePart1(String input);
+    public abstract long solvePart1(String input);
 
-    public abstract int solvePart2(String input);
+    public abstract long solvePart2(String input);
 }
